requirements.txt
asgiref==3.8.1
boto3==1.34.139
botocore==1.34.139
certifi==2024.7.4
Django==5.0.6
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.3
django-filter==24.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-nine==0.2.7
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.1
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
environs==10.3.0
jmespath==1.0.1
Markdown==3.6
marshmallow==3.21.3
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==1.26.19
